requirements.txt
python-telegram-bot==13.15
gtts
requests
beautifulsoup4
